update marital status

+++ backend/src/main/java/com/logumrh/repository/common/MaritalStatusRepository.java
@@ -3,6 +3,6 @@ package com.logumrh.repository.common;
 import com.logumrh.model.common.MaritalStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaritalStatusRepository extends JpaRepository<MaritalStatus, Integer> {
+public interface MaritalStatusRepository extends JpaRepository<MaritalStatus, Long> {
 
 }
